A raw-data file and Tree size limit is increased to 20GB. Wherever is needed the interfaces are moved from 32-bit to 64-bit integers. The mStreamRecorder interface is updated accordingly. For the moment the svn Revision keyword is used to log the aliroot version inside the raw-data tree. Should be corrected soon.

+++ RAW/mdc.cxx
@@ -51,7 +51,7 @@ int alimdcProcessEvent(void* alimdc, void* event, int isIovecArray)
   return ((AliMDC*)alimdc)->ProcessEvent(event, isIovecArray);
 }
 
-int alimdcGetTotalFileSize(void* alimdc)
+long long alimdcGetTotalFileSize(void* alimdc)
 {
 // return the total current file size
